Move GetSet into its own file
