+++ tests/helper.py
@@ -1,0 +1,9 @@
+class GetSet:
+    def __init__(self, state):
+        self.state = state
+
+    def get(self):
+        return self.state
+
+    def set(self, state):
+        self.state = state
